// src/API/useCases/ADMIN/reports/summaryReport/response.SummaryReport.ts
export interface SummaryResponse {
    errorCode: string;
    error:     string;
    data:      SummaryData;
}

export interface SummaryData {
    saletotal?:                     number;
    saletotalNegative?:             null;
    salenum?:                       number;
    neg_amt?:                       string;
    refundtotal?:                   number;
    refundnum?:                     string;
    voidtotal?:                     number;
    voidnum?:                       string;
    discountamt:                   number;
    discountqty?:                   string;
    taxqty?:                        number;
    cashpayout?:                    any[];
    total_surcharge:               string;
    total_surcharge_sales?:         number;
    netsalenum:                    number;
    neg_refs?:                      string;
    discountref?:                   string;
    netsalerefs?:                   string;
    salerounding?:                  string;
    salesroundingcount?:            string;
    saleroundingrefs:              string;
    salerefs?:                      string;
    taxes?:                         number;
    itemdeposit?:                   number;
    itemdepositrefs?:               string;
    itemdepositcount?:              number;
    itemdepositreturn?:             number;
    itemdepositreturnrefs?:         string;
    itemdepositreturncount?:        number;
    additionalcharges?:             number;
    additionalchargesresfs?:        string;
    additionalchargescount?:        number;
    refundrefs?:                    string;
    voidrefs?:                      string;
    nontaxable?:                    number;
    qtytotal?:                      string;
    nontaxrefs?:                    string;
    payouttotal?:                   number;
    payoutnum?:                     string;
    payoutrefs?:                    string;
    lottopayoutdiscount?:           string;
    lottosaletotal?:                number;
    lottosalenum?:                  string;
    lottosalerefs?:                 string;
    lottosalediscount?:             string;
    onlinepaytotal:                number;
    onlinepaynum?:                  string;
    onlinepayrefs?:                 string;
    onlinepaydiscount?:             string;
    onlinelotterytotal?:            number;
    onlinelotterynum?:              string;
    onlinelotteryrefs?:             string;
    onlinelotterydiscount?:         string;
    feeitemsaletotal:              number;
    feeitemsalenum?:                string;
    feeitemsalerefs?:               string;
    feeitemsalediscount?:           string;
    houseitemsaletotal?:            number;
    houseitemsalenum?:              string;
    houseitemsalerefs:             string;
    houseitemsalediscount?:         string;
    couponitemsaletotal?:           string;
    couponitemsalenum?:             string;
    couponitemsalerefs?:            string;
    couponitemsalediscount?:        string;
    couponpercentitemsaletotal?:    string;
    couponpercentitemsalenum?:      string;
    couponpercentitemsalerefs?:     string;
    couponpercentitemsalediscount?: string;
    pricechangetotal:              number;
    pricechangenum?:                string;
    pricechangerefs?:               string;
    cancelsaletotal?:               number;
    cancelsalenum?:                 string;
    cancelsalerefs?:                string;
    nosalestotal?:                  number;
    nosalesnum?:                    string;
    nosalesrefs?:                   string;
    tipsamounttotal?:               string;
    tipscount?:                     string;
    tipsrefs?:                      string;
    greatrating?:                   number;
    greatratingcount?:              string;
    greatratingrefs?:               string;
    okrating?:                      number;
    okratingcount?:                 string;
    okratingrefs?:                  string;
    nothappyrating?:                number;
    nothappyratingcount?:           string;
    nothappyratingrefs?:            string;
    promocouponcount?:              string;
    promocoupondisc?:               string;
    promocouponrefs?:               string;
    registerclose?:                 any[];
    cardamt?:                       number;
    cardrefs?:                      string;
    cardnum?:                       number;
    cashamt?:                       number;
    cashrefs?:                      string;
    cashnum?:                       number;
    additional_tender?:             any[];
    totaltakings?:                  number;
    profit?:                        number;
    cost?:                          number;
    refs?:                          string;
    additionalchargeslisting?: {
      [key: string]: {
      labelname: string;
      total: number;
      addnlcount: string;
      addnlchargesrefs?: string;
    };
  };
}